+++ lib/pages/settings/definitions_list_page.dart
@@ -1,12 +1,10 @@
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:intersperse/intersperse.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/app_bar/title_app_bar.dart';
+import 'package:lotti/widgets/app_bar/definitions_list_app_bar.dart';
 import 'package:lotti/widgets/settings/settings_card.dart';
-import 'package:material_floating_search_bar/material_floating_search_bar.dart';
 
 const double iconSize = 24;
 
@@ -38,54 +36,10 @@ class _DefinitionsListPageState<T> extends State<DefinitionsListPage<T>> {
     super.initState();
   }
 
-  Widget buildFloatingSearchBar() {
-    final isPortrait =
-        MediaQuery.of(context).orientation == Orientation.portrait;
-
-    final portraitWidth = MediaQuery.of(context).size.width * 0.5;
-
-    return Theme(
-      data: ThemeData(
-        brightness: styleConfig().keyboardAppearance,
-      ),
-      child: FloatingSearchBar(
-        clearQueryOnClose: false,
-        automaticallyImplyBackButton: false,
-        hint: AppLocalizations.of(context)!.settingsMeasurablesSearchHint,
-        scrollPadding: const EdgeInsets.only(top: 16, bottom: 56),
-        transitionDuration: const Duration(milliseconds: 800),
-        transitionCurve: Curves.easeInOut,
-        backgroundColor: styleConfig().cardColor,
-        margins: const EdgeInsets.only(top: 8),
-        queryStyle: const TextStyle(
-          fontFamily: mainFont,
-          fontSize: 20,
-        ),
-        hintStyle: TextStyle(
-          fontFamily: mainFont,
-          fontSize: 20,
-          color: styleConfig().secondaryTextColor,
-        ),
-        physics: const BouncingScrollPhysics(),
-        borderRadius: BorderRadius.circular(8),
-        axisAlignment: isPortrait ? 0 : -1,
-        openAxisAlignment: 0,
-        width: isPortrait ? portraitWidth : MediaQuery.of(context).size.width,
-        onQueryChanged: (query) async {
-          setState(() {
-            match = query.toLowerCase();
-          });
-        },
-        actions: [
-          FloatingSearchBarAction.searchToClear(
-            showIfClosed: false,
-          ),
-        ],
-        builder: (context, transition) {
-          return const SizedBox.shrink();
-        },
-      ),
-    );
+  Future<void> onQueryChanged(String query) async {
+    setState(() {
+      match = query.toLowerCase();
+    });
   }
 
   @override
@@ -105,7 +59,11 @@ class _DefinitionsListPageState<T> extends State<DefinitionsListPage<T>> {
             .toList();
 
         return Scaffold(
-          appBar: TitleAppBar(title: widget.title),
+          appBar: DefinitionsListAppBar(
+            title: widget.title,
+            onQueryChanged: onQueryChanged,
+            match: match,
+          ),
           backgroundColor: styleConfig().negspace,
           floatingActionButton: widget.floatingActionButton,
           body: Stack(
@@ -114,7 +72,7 @@ class _DefinitionsListPageState<T> extends State<DefinitionsListPage<T>> {
                 shrinkWrap: true,
                 padding: const EdgeInsets.only(
                   bottom: 8,
-                  top: 64,
+                  top: 8,
                 ),
                 children: intersperse(
                   const SettingsDivider(),
@@ -129,7 +87,6 @@ class _DefinitionsListPageState<T> extends State<DefinitionsListPage<T>> {
                   ),
                 ).toList(),
               ),
-              buildFloatingSearchBar(),
             ],
           ),
         );