+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simplified editor toolbar
 - Unified searchable list
 - Hide task title label when task defined
+- Floating search bar replaced in settings
+- Floating search bar replaced in journal
 
 ## [0.8.225] - 2023-01-06
 ### Changed: